comments corrected
