+++ src/test/java/praktikum/bun/GetPriceBunTest.java
@@ -1,4 +1,4 @@
-package praktikum.bun.test.suit;
+package praktikum.bun;
 
 import org.junit.Test;
 import praktikum.Bun;
@@ -6,6 +6,14 @@ import praktikum.Bun;
 import static org.junit.Assert.assertEquals;
 
 public class GetPriceBunTest {
+
+    @Test
+    public void bunGetNameTest() {
+        String name = "umwelt";
+        Bun testBun = new Bun(name, 770);
+        assertEquals(name, testBun.getName());
+    }
+
     @Test
     public void bunGetPriceTest() {
         float price = 770;
